Don't affix footer to the bottom of the window
With the footer's fixed position, it's possible for it to intersect and
overlap with the text area for pasting in code. The footer then sits on
top of the text area and looks bad. Instead, leave the footer at the
bottom of the page rather than the bottom of the window itself.

Signed-off-by: David Celis <me@davidcel.is>